Move tests into separate folders
